Fitting Count Regression Models with Latent Covariates

This function is the main function of the package and can be used to estimate latent variable count regression models in one or multiple group(s).

countreg(
  forml,
  lv = NULL,
  group = NULL,
  data,
  family,
  silent = FALSE,
  se = TRUE
)

Arguments

forml
an object of class for formula (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under Details.

lv
Definition of the latent variables.

group
A group variable

data
a data frame

family
Poisson or negative binomial

silent
Should informations about the estimation process be suppressed?

se
Should standard errors be computed? (Can take a while for complex models)

Value

An object of type lavacreg. Use summary(object) to print results containing parameter estimates and their standard errors.

Examples

fit <- countreg(forml = 'dv ~ z11', data = example01, family = 'poisson')

#> Fitting the model...done. Took: 0.6 s
#> Computing standard errors...done. Took: 0.2 s

summary(fit)

#>   lhs op rhs    dest type group par_free        par          SE
#> 1      %   w  groupw <NA>     1        1  6.7691671 0.033891754
#> 2  dv  ~   1 regcoef <NA>     1        2  2.7590148 0.014636007
#> 3  dv  ~ z11 regcoef <NA>     1        3 -0.1376593 0.008095389
#> 4 z11  ~   1       z mean     1        4  1.5792611 0.041792618
#> 5 z11 ~~ z11       z  var     1        5  1.5213083 0.072912680
#> 6  dv ~~  dv  sigmaw size     1        0  0.0000000          NA

# \donttest{
fit <- countreg(forml = 'dv ~ eta1 + z11 + z21', 
              lv = list(eta1=c('z41', 'z42', 'z43')),
              group = 'treat',
               data = example01,
             family = 'poisson')

#> Computing starting values...done. Took: 13.3 s
#> Fitting the model...done. Took: 56.2 s
#> Computing standard errors...done. Took: 74.8 s

summary(fit)

#>      lhs op  rhs    dest     type group par_free         par         SE
#> 1  treat  %    w  groupw     <NA>     1        1  6.05912137 0.04833685
#> 2     dv  ~    1 regcoef     <NA>     1        2  2.30528023 0.08445265
#> 3     dv  ~  z11 regcoef     <NA>     1        3 -0.09874118 0.01326205
#> 4     dv  ~  z21 regcoef     <NA>     1        4  0.10447775 0.01748868
#> 5     dv  ~ eta1 regcoef     <NA>     1        5 -0.08845632 0.01507367
#> 6    z41  ~    1      mm       nu     1        0  0.00000000         NA
#> 7   eta1 =~  z41      mm   lambda     1        0  1.00000000         NA
#> 8    z42  ~    1      mm       nu     1        6 -0.05920591 0.11097849
#> 9   eta1 =~  z42      mm   lambda     1        7  1.26332157 0.05464126
#> 10   z43  ~    1      mm       nu     1        8 -0.39775807 0.11737796
#> 11  eta1 =~  z43      mm   lambda     1        9  1.32903552 0.05934484
#> 12  eta1  ~    1 lv_grid     mean     1       10  1.57899293 0.07983147
#> 13  eta1 ~~ eta1 lv_grid      var     1       11  1.91178153 0.19838243
#> 14   z11  ~    1       z     mean     1       12  1.59490764 0.06259526
#> 15   z11 ~~  z11       z      var     1       13  1.68065470 0.11466445
#> 16   z21  ~    1       z     mean     1       14  3.91040943 0.04709782
#> 17   z21 ~~  z21       z      var     1       15  0.95049392 0.06499231
#> 18   z11 ~~  z21       z      cov     1       16 -0.50270092 0.06566092
#> 19  eta1 ~~  z11       z cov_z_lv     1       17  0.46661547 0.09907566
#> 20  eta1 ~~  z21       z cov_z_lv     1       18 -0.26976916 0.07399080
#> 21    dv ~~   dv  sigmaw     size     1        0  0.00000000         NA
#> 22   z41 ~~  z41  sigmaw     veps     1       19  1.51636705 0.13350153
#> 23   z42 ~~  z42  sigmaw     veps     1       20  1.45057967 0.15926675
#> 24   z43 ~~  z43  sigmaw     veps     1       21  1.46747715 0.16706319
#> 25 treat  %    w  groupw     <NA>     2       22  6.09357092 0.04751139
#> 26    dv  ~    1 regcoef     <NA>     2       23  2.47040338 0.08111459
#> 27    dv  ~  z11 regcoef     <NA>     2       24 -0.07936100 0.01316611
#> 28    dv  ~  z21 regcoef     <NA>     2       25  0.08603090 0.01657661
#> 29    dv  ~ eta1 regcoef     <NA>     2       26 -0.03120840 0.01166398
#> 30   z41  ~    1      mm       nu     2        0  0.00000000         NA
#> 31  eta1 =~  z41      mm   lambda     2        0  1.00000000         NA
#> 32   z42  ~    1      mm       nu     2       27 -0.05920591 0.11097849
#> 33  eta1 =~  z42      mm   lambda     2       28  1.26332157 0.05464126
#> 34   z43  ~    1      mm       nu     2       29 -0.39775807 0.11737796
#> 35  eta1 =~  z43      mm   lambda     2       30  1.32903552 0.05934484
#> 36  eta1  ~    1 lv_grid     mean     2       31  1.64358864 0.07421284
#> 37  eta1 ~~ eta1 lv_grid      var     2       32  2.21627193 0.24496354
#> 38   z11  ~    1       z     mean     2       33  1.54988336 0.05475422
#> 39   z11 ~~  z11       z      var     2       34  1.38395545 0.09488959
#> 40   z21  ~    1       z     mean     2       35  4.00197663 0.04304441
#> 41   z21 ~~  z21       z      var     2       36  0.84138356 0.05724378
#> 42   z11 ~~  z21       z      cov     2       37 -0.48033267 0.05728611
#> 43  eta1 ~~  z11       z cov_z_lv     2       38  0.66184701 0.10469065
#> 44  eta1 ~~  z21       z cov_z_lv     2       39 -0.40334532 0.07935062
#> 45    dv ~~   dv  sigmaw     size     2        0  0.00000000         NA
#> 46   z41 ~~  z41  sigmaw     veps     2       40  1.33561275 0.11782273
#> 47   z42 ~~  z42  sigmaw     veps     2       41  1.53723790 0.15016451
#> 48   z43 ~~  z43  sigmaw     veps     2       42  1.09534083 0.17179259

# }